File moved to Cryptography -> LogonCommDefines.h
